final, falta revisar con documentacion

+++ src/main/java/com/esliceu/maze/services/ResetService.java
@@ -25,9 +25,10 @@ public class ResetService {
 
     public void resetGame(String username) {
         User user = userDAO.getUserByUsername(username);
+        if (user.getRoomId() == null) return;
         Room actualRoom = roomDAO.getRoomById(user.getRoomId());
         Map map = mapDAO.getMapById(actualRoom.getMapId());
-        userDAO.resetUser(null, 0, null, null, null, username, null, "disconect");
+        userDAO.resetUser(null, 0, null, null, null, username, null);
         userRoomsDAO.deleteUserRoomsByUserIdAndMapId(user.getId(), map.getId());
     }
 }
